pi-release/ham_hat_control_center/requirements.txt
pyserial>=3.5
